feat: GSUB vert substitution + vertical advance
Add a GSUB lookup walker that resolves Lookup Type 1 (single
substitution), unwrapping Type 7 (extension) lookups, exposed as:

- Gsub.getSingleSubstitutions(featureTag) -> Map<glyph, substitute>
- SFNT.getSubstituteGlyphIndex(glyphIndex, featureTag)

This yields a font's real vert/vrt2 vertical-writing variants,
replacing rotation/heuristic approximations downstream.

Add vertical advance metrics:

- Glyph.advanceHeight / topSideBearing (populated from vmtx)
- SFNT.getAdvanceHeight(text, fontSize), mirroring getAdvanceWidth,
  falling back to defaultAdvanceHeight (ascender + |descender|)
  when vmtx is absent

Also fix Vmtx.readMetrics() reading topSideBearing as uint8 (1 byte)
instead of int16 (2 bytes), which misaligned the whole table and
disagreed with Vmtx.from().

+++ src/sfnt/Vmtx.ts
@@ -38,7 +38,7 @@ export class Vmtx extends SFNTTable {
       }
       return {
         advanceHeight,
-        topSideBearing: reader.readUint8(),
+        topSideBearing: reader.readInt16(),
       }
     })
   }
